Move function to address helper class
Reduce the code in the mutation file and move code related to address update and validation to the addresshelper.js file.
