+++ src/resolvers/helper/addressHelper.js
@@ -1,5 +1,5 @@
 const countries = require("countryjs");
-const {validateRequiredField} = require("./objectHelper");
+const {validateRequiredField, copyValueToObjectIfDefined} = require("./objectHelper");
 
 const throwExceptionIfProvinceDoesNotBelongToCountry = (country, province) =>{
     var states = countries.states(country);
@@ -45,8 +45,37 @@ const getNewAddressFromArgs = (args) =>{
   return args.address;
 };
 
+function updateExistingAddress(args){
+    var updateAddressData = {
+        streetAddress: copyValueToObjectIfDefined(args.address.streetAddress),
+        city: copyValueToObjectIfDefined(args.address.city),
+        postalCode: copyValueToObjectIfDefined(args.address.postalCode)
+    };
+    if (typeof args.address.country !== "undefined"){
+        var selectedCountry = args.address.country.value;
+        throwExceptionIfCountryIsDefinedButNotProvince(selectedCountry, args.address.province);
+        throwExceptionIfProvinceDoesNotBelongToCountry(selectedCountry, args.address.province);
+        updateAddressData.country = selectedCountry;
+        updateAddressData.province = args.address.province;
+    }
+    return { update: updateAddressData };
+}
+
+
+function updateOrCreateAddressOnProfile(args, profile){
+    if(typeof args.address === "undefined"){
+        return;
+    }
+    if (profile.address.id !== null){
+        return updateExistingAddress(args);
+    }
+    var newAddress = getNewAddressFromArgs(args);
+    if(newAddress != null) {
+        return { create:newAddress };
+    }
+}
+
 module.exports ={
-  throwExceptionIfProvinceDoesNotBelongToCountry,
-  throwExceptionIfCountryIsDefinedButNotProvince,
-  getNewAddressFromArgs
+  getNewAddressFromArgs,
+  updateOrCreateAddressOnProfile
 };
